+++ Classes/Private/Manager/OCTManager.m
@@ -10,6 +10,8 @@
 
 #import "OCTManager.h"
 #import "OCTTox.h"
+#import "OCTToxEncryptSave.h"
+#import "OCTToxEncryptSaveConstants.h"
 #import "OCTManagerConfiguration.h"
 #import "OCTSubmanagerAvatars+Private.h"
 #import "OCTSubmanagerBootstrap+Private.h"
@@ -41,6 +43,8 @@
 
 @property (strong, atomic) NSNotificationCenter *notificationCenter;
 
+@property (strong, nonatomic) OCTToxEncryptSave *encryptSave;
+
 @end
 
 @implementation OCTManager
@@ -49,62 +53,43 @@
 
 - (instancetype)initWithConfiguration:(OCTManagerConfiguration *)configuration error:(NSError **)error
 {
-    return [self initWithConfiguration:configuration loadToxSaveFilePath:nil error:error];
-}
-
-- (instancetype)initWithConfiguration:(OCTManagerConfiguration *)configuration
-                  loadToxSaveFilePath:(NSString *)toxSaveFilePath
-                                error:(NSError **)error
-{
     self = [super init];
 
     if (! self) {
         return nil;
     }
 
-    self.notificationCenter = [[NSNotificationCenter alloc] init];
-
     [self validateConfiguration:configuration];
-
-    NSString *savedDataPath = configuration.fileStorage.pathForToxSaveFile;
-
-    if (toxSaveFilePath && [[NSFileManager defaultManager] fileExistsAtPath:toxSaveFilePath]) {
-        [[NSFileManager defaultManager] moveItemAtPath:toxSaveFilePath toPath:savedDataPath error:nil];
-    }
-
     _configuration = [configuration copy];
 
-    NSData *savedData = nil;
-
-    if ([[NSFileManager defaultManager] fileExistsAtPath:savedDataPath]) {
-        savedData = [NSData dataWithContentsOfFile:savedDataPath];
-    }
-
-    _tox = [[OCTTox alloc] initWithOptions:configuration.options savedData:savedData error:error];
-
-    if (! _tox) {
+    if (! [self importToxSaveIfNeeded]) {
+        [self fillError:error withInitErrorCode:OCTManagerInitErrorCannotImportToxSave];
         return nil;
     }
 
-    _tox.delegate = self;
-    [_tox start];
+    NSData *savedData = [self getSavedData];
 
-    if (! savedData) {
-        [self saveTox];
+    if (! [self createEncryptSaveWithPassphrase:_configuration.passphrase toxData:savedData]) {
+        [self fillError:error withInitErrorCode:OCTManagerInitErrorPassphraseFailed];
+        return nil;
+    }
+    // no need to keep passphrase in memory.
+    _configuration.passphrase = nil;
+
+    BOOL wasDecryptError = NO;
+    savedData = [self decryptSavedDataIfNeeded:savedData error:error wasDecryptError:&wasDecryptError];
+
+    if (wasDecryptError) {
+        return nil;
     }
 
-    _realmManager = [[OCTRealmManager alloc] initWithDatabasePath:configuration.fileStorage.pathForDatabase];
+    if (! [self createToxWithSavedData:savedData error:error]) {
+        return nil;
+    }
 
-    _avatars = [self createSubmanagerWithClass:[OCTSubmanagerAvatars class]];
-    _bootstrap = [self createSubmanagerWithClass:[OCTSubmanagerBootstrap class]];
-    _chats = [self createSubmanagerWithClass:[OCTSubmanagerChats class]];
-    _dns = [self createSubmanagerWithClass:[OCTSubmanagerDNS class]];
-    _files = [self createSubmanagerWithClass:[OCTSubmanagerFiles class]];
-    _friends = [self createSubmanagerWithClass:[OCTSubmanagerFriends class]];
-    _objects = [self createSubmanagerWithClass:[OCTSubmanagerObjects class]];
-    _user = [self createSubmanagerWithClass:[OCTSubmanagerUser class]];
-
-    _toxSaveFileLock = [NSObject new];
+    [self createNotificationCenter];
+    [self createRealmManager];
+    [self createSubmanagers];
 
     return self;
 }
@@ -135,6 +120,16 @@
 
         return tempPath;
     }
+}
+
+- (BOOL)changePassphrase:(NSString *)passphrase
+{
+    // Passing nil as tox data as we are setting new passphrase.
+    BOOL result = [self createEncryptSaveWithPassphrase:passphrase toxData:nil];
+
+    [self saveTox];
+
+    return result;
 }
 
 #pragma mark -  OCTSubmanagerDataSource
@@ -171,10 +166,22 @@
 
 #pragma mark -  Private
 
+- (BOOL)createEncryptSaveWithPassphrase:(NSString *)passphrase toxData:(NSData *)toxData
+{
+    @synchronized(self.toxSaveFileLock) {
+        if (passphrase) {
+            self.encryptSave = [[OCTToxEncryptSave alloc] initWithPassphrase:passphrase toxData:toxData error:nil];
+            return (self.encryptSave != 0);
+        }
+        else {
+            self.encryptSave = nil;
+            return YES;
+        }
+    }
+}
+
 - (void)validateConfiguration:(OCTManagerConfiguration *)configuration
 {
-    NSParameterAssert(configuration.settingsStorage);
-
     NSParameterAssert(configuration.fileStorage);
     NSParameterAssert(configuration.fileStorage.pathForDownloadedFilesDirectory);
     NSParameterAssert(configuration.fileStorage.pathForUploadedFilesDirectory);
@@ -182,6 +189,138 @@
     NSParameterAssert(configuration.fileStorage.pathForAvatarsDirectory);
 
     NSParameterAssert(configuration.options);
+}
+
+- (void)createNotificationCenter
+{
+    _notificationCenter = [[NSNotificationCenter alloc] init];
+}
+
+- (BOOL)importToxSaveIfNeeded
+{
+    NSFileManager *fileManager = [NSFileManager defaultManager];
+
+    if (_configuration.importToxSaveFromPath && [fileManager fileExistsAtPath:_configuration.importToxSaveFromPath]) {
+        return [fileManager copyItemAtPath:_configuration.importToxSaveFromPath
+                                    toPath:_configuration.fileStorage.pathForToxSaveFile
+                                     error:nil];
+    }
+
+    return YES;
+}
+
+- (NSData *)getSavedData
+{
+    NSString *savedDataPath = _configuration.fileStorage.pathForToxSaveFile;
+
+    return [[NSFileManager defaultManager] fileExistsAtPath:savedDataPath] ?
+           [NSData dataWithContentsOfFile : savedDataPath] :
+           nil;
+}
+
+- (NSData *)decryptSavedDataIfNeeded:(NSData *)data error:(NSError **)error wasDecryptError:(BOOL *)wasDecryptError
+{
+    if (! data || ! _encryptSave) {
+        return data;
+    }
+
+    NSError *decryptError = nil;
+
+    NSData *result = [_encryptSave decryptData:data error:&decryptError];
+
+    if (result) {
+        return result;
+    }
+
+    *wasDecryptError = YES;
+    OCTToxEncryptSaveDecryptionError code = decryptError.code;
+
+    switch (code) {
+        case OCTToxEncryptSaveDecryptionErrorNone:
+            break;
+        case OCTToxEncryptSaveDecryptionErrorNull:
+            [self fillError:error withInitErrorCode:OCTManagerInitErrorDecryptNull];
+            break;
+        case OCTToxEncryptSaveDecryptionErrorBadFormat:
+            [self fillError:error withInitErrorCode:OCTManagerInitErrorDecryptBadFormat];
+            break;
+        case OCTToxEncryptSaveDecryptionErrorFailed:
+            [self fillError:error withInitErrorCode:OCTManagerInitErrorDecryptFailed];
+            break;
+    }
+
+    return nil;
+}
+
+- (BOOL)createToxWithSavedData:(NSData *)savedData error:(NSError **)error
+{
+    NSError *toxError = nil;
+
+    _tox = [[OCTTox alloc] initWithOptions:_configuration.options savedData:savedData error:&toxError];
+
+    if (_tox) {
+        _toxSaveFileLock = [NSObject new];
+        _tox.delegate = self;
+        [_tox start];
+
+        if (! savedData) {
+            // Tox was created for the first time, save it.
+            [self saveTox];
+        }
+
+        return YES;
+    }
+
+    OCTToxErrorInitCode code = toxError.code;
+
+    switch (code) {
+        case OCTToxErrorInitCodeUnknown:
+            [self fillError:error withInitErrorCode:OCTManagerInitErrorCreateToxUnknown];
+            break;
+        case OCTToxErrorInitCodeMemoryError:
+            [self fillError:error withInitErrorCode:OCTManagerInitErrorCreateToxMemoryError];
+            break;
+        case OCTToxErrorInitCodePortAlloc:
+            [self fillError:error withInitErrorCode:OCTManagerInitErrorCreateToxPortAlloc];
+            break;
+        case OCTToxErrorInitCodeProxyBadType:
+            [self fillError:error withInitErrorCode:OCTManagerInitErrorCreateToxProxyBadType];
+            break;
+        case OCTToxErrorInitCodeProxyBadHost:
+            [self fillError:error withInitErrorCode:OCTManagerInitErrorCreateToxProxyBadHost];
+            break;
+        case OCTToxErrorInitCodeProxyBadPort:
+            [self fillError:error withInitErrorCode:OCTManagerInitErrorCreateToxProxyBadPort];
+            break;
+        case OCTToxErrorInitCodeProxyNotFound:
+            [self fillError:error withInitErrorCode:OCTManagerInitErrorCreateToxProxyNotFound];
+            break;
+        case OCTToxErrorInitCodeEncrypted:
+            [self fillError:error withInitErrorCode:OCTManagerInitErrorCreateToxEncrypted];
+            break;
+        case OCTToxErrorInitCodeLoadBadFormat:
+            [self fillError:error withInitErrorCode:OCTManagerInitErrorCreateToxBadFormat];
+            break;
+    }
+
+    return NO;
+}
+
+- (void)createRealmManager
+{
+    _realmManager = [[OCTRealmManager alloc] initWithDatabasePath:_configuration.fileStorage.pathForDatabase];
+}
+
+- (void)createSubmanagers
+{
+    _avatars = [self createSubmanagerWithClass:[OCTSubmanagerAvatars class]];
+    _bootstrap = [self createSubmanagerWithClass:[OCTSubmanagerBootstrap class]];
+    _chats = [self createSubmanagerWithClass:[OCTSubmanagerChats class]];
+    _dns = [self createSubmanagerWithClass:[OCTSubmanagerDNS class]];
+    _files = [self createSubmanagerWithClass:[OCTSubmanagerFiles class]];
+    _friends = [self createSubmanagerWithClass:[OCTSubmanagerFriends class]];
+    _objects = [self createSubmanagerWithClass:[OCTSubmanagerObjects class]];
+    _user = [self createSubmanagerWithClass:[OCTSubmanagerUser class]];
 }
 
 - (id<OCTSubmanagerProtocol>)createSubmanagerWithClass:(Class)class
@@ -194,6 +333,67 @@
     }
 
     return submanager;
+}
+
+- (BOOL)fillError:(NSError **)error withInitErrorCode:(OCTManagerInitError)code
+{
+    if (! error) {
+        return NO;
+    }
+
+    NSString *failureReason = nil;
+
+    switch (code) {
+        case OCTManagerInitErrorPassphraseFailed:
+            failureReason = @"Cannot create symmetric key from given passphrase.";
+            break;
+        case OCTManagerInitErrorCannotImportToxSave:
+            failureReason = @"Cannot copy tox save at `importToxSaveFromPath` path.";
+            break;
+        case OCTManagerInitErrorDecryptNull:
+            failureReason = @"Cannot decrypt tox save file. Some input data was empty.";
+            break;
+        case OCTManagerInitErrorDecryptBadFormat:
+            failureReason = @"Cannot decrypt tox save file. Data has bad format.";
+            break;
+        case OCTManagerInitErrorDecryptFailed:
+            failureReason = @"Cannot decrypt tox save file. The encrypted byte array could not be decrypted. Either the data was corrupt or the password/key was incorrect.";
+            break;
+        case OCTManagerInitErrorCreateToxUnknown:
+            failureReason = @"Cannot create tox. Unknown error occurred.";
+            break;
+        case OCTManagerInitErrorCreateToxMemoryError:
+            failureReason = @"Cannot create tox. Was unable to allocate enough memory to store the internal structures for the Tox object.";
+            break;
+        case OCTManagerInitErrorCreateToxPortAlloc:
+            failureReason = @"Cannot create tox. Was unable to bind to a port.";
+            break;
+        case OCTManagerInitErrorCreateToxProxyBadType:
+            failureReason = @"Cannot create tox. Proxy type was invalid.";
+            break;
+        case OCTManagerInitErrorCreateToxProxyBadHost:
+            failureReason = @"Cannot create tox. proxyAddress had an invalid format or was nil (while proxyType was set).";
+            break;
+        case OCTManagerInitErrorCreateToxProxyBadPort:
+            failureReason = @"Cannot create tox. Proxy port was invalid.";
+            break;
+        case OCTManagerInitErrorCreateToxProxyNotFound:
+            failureReason = @"Cannot create tox. The proxy host passed could not be resolved.";
+            break;
+        case OCTManagerInitErrorCreateToxEncrypted:
+            failureReason = @"Cannot create tox. The saved data to be loaded contained an encrypted save.";
+            break;
+        case OCTManagerInitErrorCreateToxBadFormat:
+            failureReason = @"Cannot create tox. Data has bad format.";
+            break;
+    }
+
+    *error = [NSError errorWithDomain:kOCTManagerErrorDomain code:code userInfo:@{
+                  NSLocalizedDescriptionKey : @"Cannot create OCTManager",
+                  NSLocalizedFailureReasonErrorKey : failureReason
+              }];
+
+    return YES;
 }
 
 - (BOOL)respondsToSelector:(SEL)aSelector
@@ -239,13 +439,7 @@
 - (void)saveTox
 {
     @synchronized(self.toxSaveFileLock) {
-        NSString *savedDataPath = self.configuration.fileStorage.pathForToxSaveFile;
-
-        NSData *data = [self.tox save];
-
-        NSError *error;
-
-        if (! [data writeToFile:savedDataPath options:NSDataWritingAtomic error:&error]) {
+        void (^throwException)(NSError *) = ^(NSError *error) {
             NSDictionary *userInfo = nil;
 
             if (error) {
@@ -253,11 +447,35 @@
             }
 
             @throw [NSException exceptionWithName:@"saveToxException" reason:error.debugDescription userInfo:userInfo];
+        };
+
+        NSData *data = [self.tox save];
+
+        NSError *error;
+
+        if (self.encryptSave) {
+            data = [self.encryptSave encryptData:data error:&error];
+
+            if (! data) {
+                throwException(error);
+            }
+        }
+
+        if (! [data writeToFile:self.configuration.fileStorage.pathForToxSaveFile options:NSDataWritingAtomic error:&error]) {
+            throwException(error);
         }
     }
 }
 
 #pragma mark -  Deprecated
+
+- (instancetype)initWithConfiguration:(OCTManagerConfiguration *)configuration
+                  loadToxSaveFilePath:(NSString *)toxSaveFilePath
+                                error:(NSError **)error
+{
+    configuration.importToxSaveFromPath = toxSaveFilePath;
+    return [self initWithConfiguration:configuration error:error];
+}
 
 - (instancetype)initWithConfiguration:(OCTManagerConfiguration *)configuration
 {