+++ Tests/OCTManagerConfigurationTests.m
@@ -46,6 +46,8 @@
     configuration.options.proxyHost = @"proxy.address";
     configuration.options.proxyPort = 999;
     configuration.options.tcpPort = 777;
+    configuration.importToxSaveFromPath = @"save.tox";
+    configuration.passphrase = @"p@s$";
 
     OCTManagerConfiguration *c2 = [configuration copy];
 
@@ -57,6 +59,8 @@
     XCTAssertEqualObjects(configuration.options.proxyHost, c2.options.proxyHost);
     XCTAssertEqual(configuration.options.proxyPort, c2.options.proxyPort);
     XCTAssertEqual(configuration.options.tcpPort, c2.options.tcpPort);
+    XCTAssertEqualObjects(configuration.importToxSaveFromPath, c2.importToxSaveFromPath);
+    XCTAssertEqualObjects(configuration.passphrase, c2.passphrase);
 }
 
 @end